requirement.txt
asgiref==3.3.1
astroid==2.5.3
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-extensions==3.1.3
django-rest-auth==0.9.5
djangorestframework==3.12.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pylint==2.7.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
